trunk/software/caarray-ejb.jar/src/main/java/gov/nih/nci/caarray/services/arraydesign/ArrayDesignDetailsServiceBean.java
//============================================================================
// Copyright 5AM Solutions Inc, Yale University
//
// Distributed under the OSI-approved BSD 3-Clause License.
// See https://github.com/NCIP/caarray/blob/nimblegen-parser/LICENSE for details.
//============================================================================
package gov.nih.nci.caarray.services.arraydesign;

import gov.nih.nci.caarray.dao.CaArrayDaoFactory;
import gov.nih.nci.caarray.domain.array.ArrayDesign;
import gov.nih.nci.caarray.domain.array.ArrayDesignDetails;
import gov.nih.nci.caarray.services.AuthorizationInterceptor;
import gov.nih.nci.caarray.services.EntityConfiguringInterceptor;
import gov.nih.nci.caarray.services.HibernateSessionInterceptor;

import javax.annotation.security.PermitAll;
import javax.ejb.Remote;
import javax.ejb.Stateless;
import javax.interceptor.Interceptors;

/**
 * Implementation of the remote API array design detail retrieval subsystem.
 */
@Stateless
@Remote(ArrayDesignDetailsService.class)
@PermitAll
@Interceptors({ AuthorizationInterceptor.class, HibernateSessionInterceptor.class, EntityConfiguringInterceptor.class })
public class ArrayDesignDetailsServiceBean implements ArrayDesignDetailsService {

    private CaArrayDaoFactory daoFactory = CaArrayDaoFactory.INSTANCE;

    /**
     * {@inheritDoc}
     */
    public ArrayDesignDetails getDesignDetails(ArrayDesign design) {
        ArrayDesign retrievedDesign = getDaoFactory().getArrayDao().getArrayDesign(design.getId());
        if (retrievedDesign != null) {
            return retrievedDesign.getDesignDetails();
        } else {
            return null;
        }
    }

    CaArrayDaoFactory getDaoFactory() {
        return daoFactory;
    }

    void setDaoFactory(CaArrayDaoFactory daoFactory) {
        this.daoFactory = daoFactory;
    }

}
